requirements.txt
wheel
tqdm
lightning
black
ruff
hydra-core
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib
imageio
git+https://github.com/rmurai0610/diff-gaussian-rasterization-w-pose.git
git+https://github.com/hanyang-21/diff-gaussian-rasterization-mvsplat.git
timm
dacite
lpips
e3nn
plyfile
tabulate
svg.py
scikit-video
opencv-python
gradio
moviepy==1.0.3
numpy==1.26.0
diffusers>=0.31.2
transformers>=4.45.2
accelerate==0.34.0
sentencepiece>=0.2.0
decord==0.6.0